refactor: move constants to constants file and AggregatedWord type to IWord
