+++ src/constants.ts
@@ -5,6 +5,9 @@ export const TEXTBOOK_PARAMS = {
   PAGE: 'page',
 };
 
+export const SEARCH_INITIAL_GROUP = '1';
+export const SEARCH_INITIAL_PAGE = '1';
+
 export const WORDS_PER_PAGE = 20;
 export const WORDS_PER_GROUP = 600;
 export const WORDS_FOR_GAME_SPRINT = 20;
