+++ src/interfaces/IWord.ts
@@ -1,16 +1,22 @@
+import { IUserWord } from './IUserWord';
+
 export interface IWord {
-    id: string;
-    group: number;
-    page: number;
-    word: string;
-    image: string;
-    audio: string;
-    audioMeaning: string;
-    audioExample: string;
-    textMeaning: string;
-    textExample: string;
-    transcription: string;
-    wordTranslate: string;
-    textMeaningTranslate: string;
-    textExampleTranslate: string;
+  id: string;
+  group: number;
+  page: number;
+  word: string;
+  image: string;
+  audio: string;
+  audioMeaning: string;
+  audioExample: string;
+  textMeaning: string;
+  textExample: string;
+  transcription: string;
+  wordTranslate: string;
+  textMeaningTranslate: string;
+  textExampleTranslate: string;
+}
+
+export interface AggregatedWord extends IWord {
+  userWord: IUserWord | null;
 }
