+++ src/pages/textbook/helpers.ts
@@ -7,9 +7,6 @@ export const checkSearchParamsCorrect = (group: string | null, page: string | nu
   // eslint-disable-next-line
   group && page && Number(group) > 0 && Number(page) > 0 && Number(group) < 7 && Number(page) < 31;
 
-export const SEARCH_INITIAL_GROUP = '1';
-export const SEARCH_INITIAL_PAGE = '1';
-
 export const prepareNewLearnedWord = (): IUserWord => ({
   difficulty: 'easy',
   optional: {
